requirements.txt
praw
fpdf
google-generativeai
requests
python-dotenv
